pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "axiom-spec"
version = "0.1.0"
description = "A portable file format for agent minds."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "frost"},
]

dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
axiom = "axiom.cli:main"

[tool.setuptools.packages.find]
include = ["axiom*"]
